@@ -24,63 +24,65 @@
   "homepage": "https://github.com/ontodia-org/ontodia#readme",
   "scripts": {
     "build": "npm run _webpack && npm run _typings",
-    "demo": "./node_modules/.bin/webpack-dev-server --port 10444 --config webpack.demo.config.js --content-base dist/",
-    "_typings-tsc": "./node_modules/.bin/tsc --declaration --outDir dist/dts",
-    "_typings-dts-bundle": "./node_modules/.bin/dts-bundle --main dist/dts/index.d.ts --name ontodia --out ../ontodia.d.ts",
+    "demo": "./node_modules/.bin/webpack-dev-server --port 10444 --config webpack.demo.config.js",
+    "_typings-tsc": "./node_modules/.bin/tsc --declaration --emitDeclarationOnly --outDir dist/temp/dts",
+    "_typings-dts-bundle": "./node_modules/.bin/dts-bundle --main dist/temp/dts/index.d.ts --name ontodia --out ../../ontodia.d.ts",
     "_typings": "npm run _typings-tsc && npm run _typings-dts-bundle",
-    "test": "npm run _webpack && BUNDLE_PEERS=true npm run _webpack && npm run _typings",
+    "test": "npm run _test-build && npm run lint",
+    "_test-build": "npm run _webpack && BUNDLE_PEERS=true npm run _webpack && npm run _typings && npm run _webpack-examples",
     "_webpack": "./node_modules/.bin/webpack",
-    "start": "./node_modules/.bin/webpack-dev-server --port 5000 --host 0.0.0.0 --config webpack.demo.config.js --content-base dist/",
-    "lint": "./node_modules/.bin/tslint -c tslint.json 'src/**/*.ts' 'src/**/*.tsx' -t stylish"
+    "_webpack-examples": "./node_modules/.bin/webpack --config webpack.demo.config.js",
+    "start": "./node_modules/.bin/webpack-dev-server --port 5000 --host 0.0.0.0 --config webpack.demo.config.js",
+    "lint": "./node_modules/.bin/tslint -c tslint.json --project tsconfig.json --exclude '**/*.d.ts' -t stylish"
   },
   "dependencies": {
     "d3-color": "~1.0.2",
     "file-saverjs": "~1.3.6",
     "intro.js": "1.1.1",
-    "lodash": "~4.17.4",
+    "lodash": "~4.17.10",
     "n3": "~0.10.0",
     "rdf-ext": "~0.3.0",
-    "webcola": "~3.1.3",
+    "webcola": "~3.3.8",
     "whatwg-fetch": "~2.0.2"
   },
   "peerDependencies": {
-    "@types/react": "*",
     "react": "^15.0.0",
     "react-dom": "^15.0.0"
   },
   "devDependencies": {
     "@types/d3-color": "1.0.4",
-    "@types/jquery": "2.0.47",
-    "@types/lodash": "4.14.63",
-    "@types/react": "0.14.55",
-    "@types/react-dom": "0.14.19",
+    "@types/jquery": "2.0.49",
+    "@types/lodash": "4.14.108",
+    "@types/react": "15.6.9",
+    "@types/react-dom": "15.5.8",
     "canvg-fixed": "1.0.0",
-    "css-loader": "0.23.1",
+    "css-loader": "1.0.0",
     "dts-bundle": "0.7.2",
-    "es6-promise": "4.2.2",
-    "exports-loader": "0.6.3",
-    "expose-loader": "0.7.1",
-    "file-loader": "0.8.5",
-    "html-webpack-plugin": "2.22.0",
+    "es6-promise": "4.2.4",
+    "exports-loader": "0.7.0",
+    "html-webpack-plugin": "3.2.0",
     "jquery": "2.2.4",
-    "json-loader": "0.5.4",
     "jstree": "3.2.1",
-    "node-sass": "4.7.2",
+    "node-sass": "4.9.3",
     "raw-loader": "0.5.1",
     "rdf-parser-jsonld": "0.3.3",
     "rdf-parser-n3": "0.3.0",
     "rdf-parser-rdfxml": "0.3.1",
-    "react": "15.3.2",
-    "react-dom": "15.3.2",
-    "sass-loader": "4.0.2",
-    "style-loader": "0.13.1",
-    "ts-loader": "1.3.3",
-    "tslib": "1.9.0",
-    "tslint": "4.5.1",
-    "typescript": "2.3.2",
-    "url-loader": "0.5.7",
-    "webpack": "1.13.2",
-    "webpack-dev-server": "1.16.2"
+    "react": "15.6.2",
+    "react-dom": "15.6.2",
+    "sass-loader": "7.1.0",
+    "style-loader": "0.23.0",
+    "ts-loader": "5.1.1",
+    "tslib": "1.9.3",
+    "tslint": "5.11.0",
+    "typescript": "3.0.3",
+    "url-loader": "1.1.1",
+    "webpack": "4.19.0",
+    "webpack-cli": "3.1.0",
+    "webpack-dev-server": "3.1.8"
+  },
+  "resolutions": {
+    "@types/react": "15.6.9"
   },
   "main": "dist/ontodia.js",
   "typings": "dist/ontodia.d.ts"
