@@ -1,4 +1,4 @@
-import { LayoutData } from '../index';
+import { LayoutData, SerializedDiagram, convertToSerializedDiagram, makeSerializedDiagram } from '../index';
 
 export function onPageLoad(callback: (container: HTMLDivElement) => void) {
     document.addEventListener('DOMContentLoaded', () => {
@@ -9,20 +9,28 @@ export function onPageLoad(callback: (container: HTMLDivElement) => void) {
     });
 }
 
-export function tryLoadLayoutFromLocalStorage(): LayoutData | undefined {
+export function tryLoadLayoutFromLocalStorage(): SerializedDiagram | undefined {
     if (window.location.hash.length > 1) {
         try {
             const key = window.location.hash.substring(1);
             const unparsedLayout = localStorage.getItem(key);
-            return unparsedLayout && JSON.parse(unparsedLayout);
+            const entry = unparsedLayout && JSON.parse(unparsedLayout);
+
+            // backward compatibility test. If we encounder old diagram,
+            // wrap it into Diagram interface, jsonld - pass through
+            if (entry['@context']) {
+                return entry;
+            } else {
+                return convertToSerializedDiagram({layoutData: entry, linkTypeOptions: []});
+            }
         } catch (e) { /* ignore */ }
     }
     return undefined;
 }
 
-export function saveLayoutToLocalStorage(layout: LayoutData): string {
+export function saveLayoutToLocalStorage(diagram: SerializedDiagram): string {
     const randomKey = Math.floor((1 + Math.random()) * 0x10000000000)
         .toString(16).substring(1);
-    localStorage.setItem(randomKey, JSON.stringify(layout));
+    localStorage.setItem(randomKey, JSON.stringify(diagram));
     return randomKey;
 }
