@@ -12,7 +12,7 @@ function onWorkspaceMounted(workspace: Workspace) {
 
     const model = workspace.getModel();
     const sparqlDataProvider = new SparqlDataProvider({
-        endpointUrl: '/sparql-endpoint',
+        endpointUrl: '/sparql',
         queryMethod: SparqlQueryMethod.GET
     }, OWLStatsSettings);
     const graphBuilder = new SparqlGraphBuilder(sparqlDataProvider);
@@ -25,16 +25,16 @@ function onWorkspaceMounted(workspace: Workspace) {
             ?propValue2 ?propType2 ?inst .
         } WHERE {
             BIND (<http://collection.britishmuseum.org/id/object/JCF8939> as ?inst)
-            ?inst rdf:type ?class.	
+            ?inst rdf:type ?class.
             OPTIONAL {?inst rdfs:label ?label}
-            OPTIONAL {?inst ?propType1 ?propValue1.  FILTER(isURI(?propValue1)). }  	
-            OPTIONAL {?propValue2 ?propType2 ?inst.  FILTER(isURI(?propValue2)). }  
+            OPTIONAL {?inst ?propType1 ?propValue1.  FILTER(isURI(?propValue1)). }
+            OPTIONAL {?propValue2 ?propType2 ?inst.  FILTER(isURI(?propValue2)). }
         } LIMIT 100`,
     );
     workspace.showWaitIndicatorWhile(loadingGraph);
 
-    loadingGraph.then(({layoutData, preloadedElements}) => model.importLayout({
-        layoutData,
+    loadingGraph.then(({diagram, preloadedElements}) => model.importLayout({
+        diagram,
         preloadedElements,
         dataProvider: sparqlDataProvider,
     })).then(() => {
