@@ -1,11 +1,9 @@
 import { createElement, ClassAttributes } from 'react';
 import * as ReactDOM from 'react-dom';
 
-import { Workspace, WorkspaceProps } from '../index';
+import { Workspace, WorkspaceProps, RDFDataProvider, GraphBuilder } from '../index';
 
 import { onPageLoad } from './common';
-import { RDFDataProvider } from '../ontodia/data/rdf/rdfDataProvider';
-import { GraphBuilder } from '../ontodia/data/sparql/graphBuilder';
 
 const N3Parser: any = require('rdf-parser-n3');
 const RdfXmlParser: any = require('rdf-parser-rdfxml');
@@ -14,7 +12,7 @@ const JsonLdParser: any = require('rdf-parser-jsonld');
 const EXAMPLE = `@prefix fts: <https://w3id.org/datafabric.cc/ontologies/fts#> .
  @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
  @prefix ex: <http://example.com/> .
-  
+
  ex:LED-1 a fts:C4_Legal_Entity_Details ;
     fts:p80_describes_company ex:LE-1 ;
     fts:p81_includes_le_reg_entity ex:LE-1-NameEntity-1157847449121,
@@ -39,19 +37,18 @@ const EXAMPLE = `@prefix fts: <https://w3id.org/datafabric.cc/ontologies/fts#> .
         fts:p82_refers_to_company ex:LE-1 ;
         fts:p76_entered_on_registry_with ex:RND-1157847449121 ;
         fts:p28_le_primary_state_registration_number "1157847449121" ;
-        fts:p29_psrn_assignment_date "25-12-2015"^^xsd:date .                            
+        fts:p29_psrn_assignment_date "25-12-2015"^^xsd:date .
 `;
 
 const DIAGRAM = `@prefix fts: <https://w3id.org/datafabric.cc/ontologies/fts#> .
  @prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
  @prefix ex: <http://example.com/> .
-  
+
  ex:LED-1 a fts:C4_Legal_Entity_Details ;
     fts:p80_describes_company ex:LE-1 ;
     fts:p81_includes_le_reg_entity ex:LE-1-NameEntity-1157847449121,
                         ex:LE-1-Address-1157847449121,
                         ex:LE-1-RegInfo-1157847449121 .`;
-
 
 function onWorkspaceMounted(workspace: Workspace) {
     if (!workspace) { return; }
@@ -75,10 +72,10 @@ function onWorkspaceMounted(workspace: Workspace) {
 
     workspace.showWaitIndicatorWhile(loadingGraph);
 
-    loadingGraph.then(({layoutData, preloadedElements}) => {
+    loadingGraph.then(({diagram, preloadedElements}) => {
         const model = workspace.getModel();
         return model.importLayout({
-            layoutData,
+            diagram,
             preloadedElements,
             dataProvider: provider,
         });
