@@ -4,13 +4,18 @@ import * as ReactDOM from 'react-dom';
 import { Workspace, WorkspaceProps, DemoDataProvider, ToolbarProps } from '../index';
 import { onPageLoad, tryLoadLayoutFromLocalStorage, saveLayoutToLocalStorage } from './common';
 
+const CLASSES = require<any>('./resources/classes.json');
+const LINK_TYPES = require<any>('./resources/linkTypes.json');
+const ELEMENTS = require<any>('./resources/elements.json');
+const LINKS  = require<any>('./resources/links.json');
+
 export interface Props extends ToolbarProps {
     onExampleClick?: () => void;
 }
 
 const CLASS_NAME = 'ontodia-toolbar';
 
-export class Toolbar extends React.Component<Props, void> {
+export class Toolbar extends React.Component<Props, {}> {
     render() {
         return (
             <div className={CLASS_NAME}>
@@ -40,15 +45,19 @@ function onWorkspaceMounted(workspace: Workspace) {
 
     const model = workspace.getModel();
 
-    const layoutData = tryLoadLayoutFromLocalStorage();
-    model.importLayout({layoutData, dataProvider: new DemoDataProvider(), validateLinks: true});
+    const diagram = tryLoadLayoutFromLocalStorage();
+    model.importLayout({
+        dataProvider: new DemoDataProvider(CLASSES, LINK_TYPES, ELEMENTS, LINKS),
+        diagram,
+        validateLinks: true,
+    });
 }
 
 const props: WorkspaceProps & React.ClassAttributes<Workspace> = {
     ref: onWorkspaceMounted,
     onSaveDiagram: workspace => {
-        const {layoutData} = workspace.getModel().exportLayout();
-        window.location.hash = saveLayoutToLocalStorage(layoutData);
+        const diagram = workspace.getModel().exportLayout();
+        window.location.hash = saveLayoutToLocalStorage(diagram);
         window.location.reload();
     },
     toolbar: <Toolbar onExampleClick={() => { alert('Example button have been pressed!'); }}/>,
