@@ -9,6 +9,8 @@ export * from './ontodia/customization/props';
 export * from './ontodia/customization/templates';
 
 export * from './ontodia/data/model';
+export { MetadataApi } from './ontodia/data/metadataApi';
+export { ValidationApi, ElementError, LinkError } from './ontodia/data/validationApi';
 export * from './ontodia/data/provider';
 export * from './ontodia/data/demo/provider';
 export { RdfNode, RdfIri, RdfLiteral, Triple } from './ontodia/data/sparql/sparqlModels';
@@ -18,21 +20,38 @@ export * from './ontodia/data/composite/composite';
 export * from './ontodia/data/sparql/sparqlDataProviderSettings';
 export * from './ontodia/data/sparql/graphBuilder';
 export * from './ontodia/data/sparql/sparqlGraphBuilder';
+export { DIAGRAM_CONTEXT_URL_V1 } from './ontodia/data/schema';
 
-export { Element, ElementEvents, Link, LinkEvents } from './ontodia/diagram/elements';
-export { LayoutData, LayoutCell, LayoutElement, LayoutLink } from './ontodia/diagram/layoutData';
+export { RestoreGeometry } from './ontodia/diagram/commands';
+export { Element, ElementEvents, Link, LinkEvents, LinkVertex, Cell } from './ontodia/diagram/elements';
 export { EmbeddedLayer } from './ontodia/diagram/embeddedLayer';
+export * from './ontodia/diagram/geometry';
+export * from './ontodia/diagram/history';
 export * from './ontodia/diagram/model';
 export * from './ontodia/diagram/view';
-export { Cell, isLinkVertex } from './ontodia/diagram/paper';
 export { PointerEvent, PointerUpEvent } from './ontodia/diagram/paperArea';
 
+export * from './ontodia/editor/asyncModel';
+export { AuthoredEntity, AuthoredEntityProps, AuthoredEntityContext } from './ontodia/editor/authoredEntity';
+export * from './ontodia/editor/authoringState';
+export {
+    EditorOptions, EditorEvents, EditorController, PropertyEditor, PropertyEditorOptions,
+} from './ontodia/editor/editorController';
+
+export {
+    LayoutData, LayoutElement, LayoutLink, SerializedDiagram,
+    convertToSerializedDiagram, makeSerializedDiagram
+} from './ontodia/editor/serializedDiagram';
+export { recursiveLayout } from './ontodia/viewUtils/layout';
+
+export { CancellationToken } from './ontodia/viewUtils/async';
 export * from './ontodia/viewUtils/events';
 
 export { PropertySuggestionParams, PropertyScore } from './ontodia/widgets/connectionsMenu';
 
 export * from './ontodia/workspace/toolbar';
 export { Workspace, WorkspaceProps, WorkspaceLanguage, renderTo } from './ontodia/workspace/workspace';
+export { WorkspaceEventHandler, WorkspaceEventKey } from './ontodia/workspace/workspaceContext';
 
 import * as InternalApi from './internalApi';
 export { InternalApi };
