@@ -1,8 +1,6 @@
 import * as React from 'react';
 import { ReactElement } from 'react';
 
-import { CrossOriginImage } from '../../viewUtils/crossOriginImage';
-
 import { TemplateProps } from '../props';
 
 const CLASS_NAME = 'ontodia-default-template';
@@ -12,8 +10,9 @@ export class DefaultElementTemplate extends React.Component<TemplateProps, {}> {
         const props = this.props;
 
         const image = props.imgUrl ? (
-            <CrossOriginImage className={`${CLASS_NAME}__thumbnail`}
-                imageProps={{src: props.imgUrl}} />
+            <div className={`${CLASS_NAME}__thumbnail`}>
+                <img src={props.imgUrl} />
+            </div>
         ) : undefined;
 
         const expander = props.isExpanded ? (
@@ -39,7 +38,8 @@ export class DefaultElementTemplate extends React.Component<TemplateProps, {}> {
                 borderColor: props.color,
             }} data-expanded={this.props.isExpanded}>
                 <div className='ontodia-default-template_type-line' title={props.label}>
-                    <div className={`${props.icon} ontodia-default-template_type-line__icon`} aria-hidden='true'>
+                    <div className='ontodia-default-template_type-line__icon' aria-hidden='true'>
+                        <img src={props.iconUrl} />
                     </div>
                     <div title={props.types} className='ontodia-default-template_type-line_text-container'>
                         <div className='ontodia-default-template_type-line_text-container__text'>
