@@ -1,11 +1,11 @@
-/** Generates random 16-digit hexadecimal string. */
-export function generate64BitID() {
-    function randomHalfDigits() {
+/** Generates random 32-digit hexadecimal string. */
+export function generate128BitID() {
+    function random32BitDigits() {
         return Math.floor((1 + Math.random()) * 0x100000000)
             .toString(16).substring(1);
     }
     // generate by half because of restricted numerical precision
-    return randomHalfDigits() + randomHalfDigits();
+    return random32BitDigits() + random32BitDigits() + random32BitDigits() + random32BitDigits();
 }
 
 /**
@@ -27,4 +27,21 @@ export function hashFnv32a(str: string, seed = 0x811c9dc5): number {
     }
     return hval >>> 0;
     /* tslint:enable:no-bitwise */
+}
+
+export function uri2name(uri: string): string {
+    const hashIndex = uri.lastIndexOf('#');
+    if (hashIndex !== -1 && hashIndex !== uri.length - 1) {
+        return uri.substring(hashIndex + 1);
+    }
+    const endsWithSlash = uri[uri.length - 1] === '/';
+    if (endsWithSlash) {
+        uri = uri.substring(0, uri.length - 1);
+    }
+
+    const lastPartStart = uri.lastIndexOf('/');
+    if (lastPartStart !== -1 && lastPartStart !== uri.length - 1) {
+        return uri.substring(lastPartStart + 1);
+    }
+    return uri;
 }
