@@ -2,13 +2,15 @@ import * as React from 'react';
 import { findDOMNode } from 'react-dom';
 import { hcl } from 'd3-color';
 
-import { Property } from '../data/model';
+import { Property, ElementTypeIri, PropertyTypeIri } from '../data/model';
 import { TemplateProps } from '../customization/props';
 import { Debouncer } from '../viewUtils/async';
 import { createStringMap } from '../viewUtils/collections';
 import { EventObserver, Unsubscribe } from '../viewUtils/events';
 import { PropTypes } from '../viewUtils/react';
+import { KeyedObserver, observeElementTypes, observeProperties } from '../viewUtils/keyedObserver';
 
+import { setElementExpanded } from './commands';
 import { Element } from './elements';
 import { formatLocalizedLabel } from './model';
 import { DiagramView, RenderingLayer } from './view';
@@ -16,7 +18,6 @@ import { DiagramView, RenderingLayer } from './view';
 export interface Props {
     view: DiagramView;
     group?: string;
-    scale: number;
     style: React.CSSProperties;
 }
 
@@ -25,36 +26,45 @@ interface BatchUpdateItem {
     node: HTMLDivElement;
 }
 
-export class ElementLayer extends React.Component<Props, void> {
+export class ElementLayer extends React.Component<Props, {}> {
     private readonly listener = new EventObserver();
 
     private batch = createStringMap<BatchUpdateItem>();
+    private renderElements = new Debouncer();
     private updateSizes = new Debouncer();
 
     private layer: HTMLDivElement;
 
     render() {
-        const {view, group, scale, style} = this.props;
+        const {view, group, style} = this.props;
         const models = view.model.elements.filter(model => model.group === group);
 
         return <div className='ontodia-element-layer'
-            ref={layer => this.layer = layer}
+            ref={this.onMount}
             style={style}>
             {models.map(model => <OverlayedElement key={model.id}
                 model={model}
                 view={view}
-                scale={scale}
                 onResize={this.updateElementSize}
                 onRender={this.updateElementSize} />)}
         </div>;
     }
 
+    private onMount = (layer: HTMLDivElement) => {
+        this.layer = layer;
+    }
+
     componentDidMount() {
         const {view} = this.props;
-        this.listener.listen(view.model.events, 'changeCells', () => this.forceUpdate());
+        this.listener.listen(view.model.events, 'changeCells', () => {
+            this.renderElements.call(this.performRendering);
+        });
         this.listener.listen(view.events, 'syncUpdate', ({layer}) => {
-            if (layer !== RenderingLayer.ElementSize) { return; }
-            this.updateSizes.runSynchronously();
+            if (layer === RenderingLayer.Element) {
+                this.renderElements.runSynchronously();
+            } else if (layer === RenderingLayer.ElementSize) {
+                this.updateSizes.runSynchronously();
+            }
         });
     }
 
@@ -64,7 +74,12 @@ export class ElementLayer extends React.Component<Props, void> {
 
     componentWillUnmount() {
         this.listener.stopListening();
+        this.renderElements.dispose();
         this.updateSizes.dispose();
+    }
+
+    private performRendering = () => {
+        this.forceUpdate();
     }
 
     private updateElementSize = (element: Element, node: HTMLDivElement) => {
@@ -87,9 +102,8 @@ export class ElementLayer extends React.Component<Props, void> {
 }
 
 interface OverlayedElementProps {
-    model: Element;
     view: DiagramView;
-    scale: number;
+    model: Element;
     onResize: (model: Element, node: HTMLDivElement) => void;
     onRender: (model: Element, node: HTMLDivElement) => void;
 }
@@ -98,54 +112,36 @@ interface OverlayedElementState {
     readonly templateProps?: TemplateProps;
 }
 
-export const ElementContextTypes = {
-    ontodiaElementContext: PropTypes.anything,
+export interface ElementContextWrapper { ontodiaElement: ElementContext; }
+export const ElementContextTypes: { [K in keyof ElementContextWrapper]: any } = {
+    ontodiaElement: PropTypes.anything,
 };
 
 export interface ElementContext {
-    view: DiagramView;
     element: Element;
-    scale: number;
 }
 
 class OverlayedElement extends React.Component<OverlayedElementProps, OverlayedElementState> {
     static childContextTypes = ElementContextTypes;
 
-    getChildContext() {
-        const ontodiaElementContext: ElementContext = {
-            view: this.props.view,
-            element: this.props.model,
-            scale: this.props.scale,
-        };
-        return {ontodiaElementContext};
-    }
-
     private readonly listener = new EventObserver();
     private disposed = false;
 
-    private typesObserver = new KeyedObserver(key => {
-        const type = this.props.view.model.getClassesById(key);
-        if (type) {
-            type.events.on('changeLabel', this.rerenderTemplate);
-            return () => type.events.off('changeLabel', this.rerenderTemplate);
-        }
-        return undefined;
-    });
-
-    private propertyObserver = new KeyedObserver(key => {
-        const property = this.props.view.model.getPropertyById(key);
-        if (property) {
-            property.events.on('changeLabel', this.rerenderTemplate);
-            return () => property.events.off('changeLabel', this.rerenderTemplate);
-        }
-        return undefined;
-    });
+    private typesObserver: KeyedObserver<ElementTypeIri>;
+    private propertiesObserver: KeyedObserver<PropertyTypeIri>;
 
     constructor(props: OverlayedElementProps) {
         super(props);
         this.state = {
             templateProps: this.templateProps(),
         };
+    }
+
+    getChildContext(): ElementContextWrapper {
+        const ontodiaElement: ElementContext = {
+            element: this.props.model,
+        };
+        return {ontodiaElement};
     }
 
     private rerenderTemplate = () => {
@@ -155,14 +151,14 @@ class OverlayedElement extends React.Component<OverlayedElementProps, OverlayedE
 
     render(): React.ReactElement<any> {
         const {model, view, onResize, onRender} = this.props;
-
-        this.typesObserver.observe(model.data.types);
-        this.propertyObserver.observe(Object.keys(model.data.properties));
+        if (model.temporary) {
+            return <div />;
+        }
 
         const template = view.getElementTemplate(model.data.types);
 
         const {x = 0, y = 0} = model.position;
-        let transform = `translate(${x}px,${y}px)`;
+        const transform = `translate(${x}px,${y}px)`;
 
         // const angle = model.get('angle') || 0;
         // if (angle) { transform += `rotate(${angle}deg)`; }
@@ -172,26 +168,42 @@ class OverlayedElement extends React.Component<OverlayedElementProps, OverlayedE
             data-element-id={model.id}
             style={{position: 'absolute', transform}}
             tabIndex={0}
+            ref={this.onMount}
             // resize element when child image loaded
-            onLoad={() => onResize(model, findDOMNode(this) as HTMLDivElement)}
-            onError={() => onResize(model, findDOMNode(this) as HTMLDivElement)}
-            onClick={e => {
-                if (e.target instanceof HTMLElement && e.target.localName === 'a') {
-                    const anchor = e.target as HTMLAnchorElement;
-                    view.onIriClick(anchor.href, model, e);
-                }
-            }}
-            onDoubleClick={e => {
-                e.preventDefault();
-                e.stopPropagation();
-                model.setExpanded(!model.isExpanded);
-            }}
-            ref={node => {
-                if (!node) { return; }
-                onRender(model, node);
-            }}>
+            onLoad={this.onLoadOrErrorEvent}
+            onError={this.onLoadOrErrorEvent}
+            onClick={this.onClick}
+            onDoubleClick={this.onDoubleClick}>
             {React.createElement(template, this.state.templateProps)}
         </div>;
+    }
+
+    private onMount = (node: HTMLDivElement | undefined) => {
+        if (!node) { return; }
+        const {onRender, model} = this.props;
+        onRender(model, node);
+    }
+
+    private onLoadOrErrorEvent = () => {
+        const {onResize, model} = this.props;
+        onResize(model, findDOMNode(this) as HTMLDivElement);
+    }
+
+    private onClick = (e: React.MouseEvent<EventTarget>) => {
+        if (e.target instanceof HTMLElement && e.target.localName === 'a') {
+            const anchor = e.target as HTMLAnchorElement;
+            const {view, model} = this.props;
+            view.onIriClick(anchor.href, model, e);
+        }
+    }
+
+    private onDoubleClick = (e: React.MouseEvent<HTMLDivElement>) => {
+        e.preventDefault();
+        e.stopPropagation();
+        const {view, model} = this.props;
+        view.model.history.execute(
+            setElementExpanded(model, !model.isExpanded)
+        );
     }
 
     componentDidMount() {
@@ -205,12 +217,19 @@ class OverlayedElement extends React.Component<OverlayedElementProps, OverlayedE
             const element = findDOMNode(this) as HTMLElement;
             if (element) { element.focus(); }
         });
+        this.typesObserver = observeElementTypes(
+            this.props.view.model, 'changeLabel', this.rerenderTemplate
+        );
+        this.propertiesObserver = observeProperties(
+            this.props.view.model, 'changeLabel', this.rerenderTemplate
+        );
+        this.observeTypes();
     }
 
     componentWillUnmount() {
         this.listener.stopListening();
         this.typesObserver.stopListening();
-        this.propertyObserver.stopListening();
+        this.propertiesObserver.stopListening();
         this.disposed = true;
     }
 
@@ -219,7 +238,14 @@ class OverlayedElement extends React.Component<OverlayedElementProps, OverlayedE
     }
 
     componentDidUpdate() {
+        this.observeTypes();
         this.props.onResize(this.props.model, findDOMNode(this) as HTMLDivElement);
+    }
+
+    private observeTypes() {
+        const {model} = this.props;
+        this.typesObserver.observe(model.data.types);
+        this.propertiesObserver.observe(Object.keys(model.data.properties) as PropertyTypeIri[]);
     }
 
     private templateProps(): TemplateProps {
@@ -232,11 +258,13 @@ class OverlayedElement extends React.Component<OverlayedElementProps, OverlayedE
         const propsAsList = this.getPropertyTable();
 
         return {
+            elementId: model.id,
+            data: model.data,
+            iri: model.iri,
             types,
             label,
             color,
-            icon,
-            iri: model.iri,
+            iconUrl: icon,
             imgUrl: model.data.image,
             isExpanded: model.isExpanded,
             props: model.data.properties,
@@ -244,13 +272,14 @@ class OverlayedElement extends React.Component<OverlayedElementProps, OverlayedE
         };
     }
 
-    private getPropertyTable(): Array<{ id: string; name: string; property: Property; }> {
+    private getPropertyTable(): Array<{ id: string; name: string; property: Property }> {
         const {model, view} = this.props;
 
         if (!model.data.properties) { return []; }
 
-        const propTable = Object.keys(model.data.properties).map(key => {
-            const property = view.model.getPropertyById(key);
+        const propertyIris = Object.keys(model.data.properties) as PropertyTypeIri[];
+        const propTable = propertyIris.map(key => {
+            const property = view.model.createProperty(key);
             const name = formatLocalizedLabel(key, property.label, view.getLanguage());
             return {
                 id: key,
@@ -270,40 +299,8 @@ class OverlayedElement extends React.Component<OverlayedElementProps, OverlayedE
     private styleFor(model: Element) {
         const {color: {h, c, l}, icon} = this.props.view.getTypeStyle(model.data.types);
         return {
-            icon: icon ? icon : 'ontodia-default-icon',
+            icon,
             color: hcl(h, c, l).toString(),
         };
-    }
-}
-
-class KeyedObserver {
-    private observedKeys = createStringMap<Unsubscribe>();
-
-    constructor(readonly subscribe: (key: string) => Unsubscribe | undefined) {}
-
-    observe(keys: string[]) {
-        const newObservedKeys = createStringMap<Unsubscribe>();
-
-        for (const key of keys) {
-            if (newObservedKeys[key]) { continue; }
-            let token = this.observedKeys[key];
-            if (!token) {
-                token = this.subscribe(key);
-            }
-            newObservedKeys[key] = token;
-        }
-
-        for (const key in this.observedKeys) {
-            if (!newObservedKeys[key]) {
-                const unsubscribe = this.observedKeys[key];
-                unsubscribe();
-            }
-        }
-
-        this.observedKeys = newObservedKeys;
-    }
-
-    stopListening() {
-        this.observe([]);
     }
 }
