@@ -1,18 +1,13 @@
 import * as React from 'react';
 
-import { Dictionary, ElementModel } from '../data/model';
-import { Paper, Cell } from './paper';
-import { Element } from './elements';
-import { ElementLayer, ElementContext, ElementContextTypes } from './elementLayer';
+import { Dictionary, ElementModel, ElementIri } from '../data/model';
+import { Paper, PaperTransform } from './paper';
+import { PaperAreaContextTypes, PaperAreaContextWrapper } from './paperArea';
+import { Element, Cell } from './elements';
+import { ElementLayer, ElementContextWrapper, ElementContextTypes } from './elementLayer';
 import { EventObserver } from '../viewUtils/events';
-import {
-    LayoutNode,
-    LayoutLink,
-    forceLayout,
-    padded,
-    removeOverlaps,
-    translateToPositiveQuadrant,
-} from '../viewUtils/layout';
+
+import { Vector, Rect } from './geometry';
 import { getContentFittingBox } from './paperArea';
 import { DiagramView } from './view';
 
@@ -24,19 +19,19 @@ export interface State {
 }
 
 export class EmbeddedLayer extends React.Component<{}, State> {
-    static contextTypes = ElementContextTypes;
+    static contextTypes = {...ElementContextTypes, ...PaperAreaContextTypes};
 
-    context: {
-        ontodiaElementContext: ElementContext;
-    };
+    context: ElementContextWrapper & PaperAreaContextWrapper;
 
     private readonly listener = new EventObserver();
+    private nestedElementListener = new EventObserver();
 
     private layerOffsetLeft = 0;
     private layerOffsetTop = 0;
 
     private isApplyingParentMove = false;
     private isNestedElementMoving = false;
+    private previousPositions: Array<{ id: string; position: Vector }> = [];
 
     constructor(props: {}) {
         super(props);
@@ -44,9 +39,16 @@ export class EmbeddedLayer extends React.Component<{}, State> {
     }
 
     componentDidMount() {
-        const {element} = this.context.ontodiaElementContext;
+        const {element} = this.context.ontodiaElement;
+        const {paperArea, view} = this.context.ontodiaPaperArea;
 
-        document.addEventListener('mouseup', this.onMouseUp);
+        this.listener.listen(view.model.events, 'changeGroupContent', ({group}) => {
+            if (group === element.id) {
+                this.listenNestedElements(this.getNestedElements());
+                const {offsetX, offsetY} = this.getOffset();
+                this.moveNestedElements(offsetX, offsetY);
+            }
+        });
 
         this.listener.listen(element.events, 'changePosition', () => {
             if (this.isNestedElementMoving) { return; }
@@ -61,8 +63,12 @@ export class EmbeddedLayer extends React.Component<{}, State> {
             this.setState({offsetX, offsetY});
         });
 
+        this.listener.listen(paperArea.events, 'pointerUp', e => {
+            this.isNestedElementMoving = false;
+        });
+
         const nestedElements = this.getNestedElements();
-        nestedElements.forEach(this.listenNestedElement);
+        this.listenNestedElements(nestedElements);
 
         if (nestedElements.length > 0) {
             const {
@@ -73,62 +79,49 @@ export class EmbeddedLayer extends React.Component<{}, State> {
             } = this.getContentFittingBox();
             this.setState({offsetX, offsetY, paperWidth, paperHeight}, () => element.redraw());
         } else {
-            this.loadEmbeddedElements();
+            element.requestGroupContent();
         }
     }
 
+    private listenNestedElements(nestedElements: ReadonlyArray<Element>) {
+        const listener = new EventObserver();
+        for (const nestedElement of nestedElements) {
+            listener.listen(nestedElement.events, 'changePosition', this.recomputeSelfBounds);
+            listener.listen(nestedElement.events, 'changeSize', this.recomputeSelfBounds);
+        }
+        this.nestedElementListener.stopListening();
+        this.nestedElementListener = listener;
+    }
+
     componentWillUnmount() {
-        document.removeEventListener('mouseup', this.onMouseUp);
         this.listener.stopListening();
+        this.nestedElementListener.stopListening();
         this.removeElements();
         this.setState({paperWidth: 0, paperHeight: 0, offsetX: 0, offsetY: 0});
     }
 
-    private onMouseUp = () => {
-        this.isNestedElementMoving = false;
-    }
-
-    private getNestedElements = () => {
-        const {view, element} = this.context.ontodiaElementContext;
+    private getNestedElements() {
+        const {element} = this.context.ontodiaElement;
+        const {view} = this.context.ontodiaPaperArea;
         return view.model.elements.filter(el => el.group === element.id);
     }
 
-    private getContentFittingBox = (): { x: number; y: number; width: number; height: number; } => {
+    private getContentFittingBox(): Rect {
         const nestedElements = this.getNestedElements();
         return getContentFittingBox(nestedElements, []);
     }
 
-    private removeElements = () => {
-        const {view} = this.context.ontodiaElementContext;
+    private removeElements() {
+        const {view} = this.context.ontodiaPaperArea;
+        const batch = view.model.history.startBatch();
         for (const element of this.getNestedElements()) {
             view.model.removeElement(element.id);
         }
+        batch.discard();
     }
 
-    private loadEmbeddedElements = () => {
-        const {view, element} = this.context.ontodiaElementContext;
-
-        view.loadEmbeddedElements(element.iri).then(models => {
-            const elements = Object.keys(models)
-                .map(key => view.model.createElement(models[key], element.id));
-
-            elements.forEach(this.listenNestedElement);
-
-            Promise.all([
-                view.model.requestElementData(elements),
-                view.model.requestLinksOfType(),
-            ]).then(() => {
-                view.performSyncUpdate();
-                this.forceLayout();
-
-                const {offsetX, offsetY} = this.getOffset();
-                this.moveNestedElements(offsetX, offsetY);
-            });
-        });
-    }
-
-    private getOffset = (): { offsetX: number; offsetY: number; } => {
-        const {element} = this.context.ontodiaElementContext;
+    private getOffset(): { offsetX: number; offsetY: number } {
+        const {element} = this.context.ontodiaElement;
         const {x: elementX, y: elementY} = element.position;
 
         const offsetX = elementX + this.layerOffsetLeft;
@@ -151,15 +144,10 @@ export class EmbeddedLayer extends React.Component<{}, State> {
         }
     }
 
-    private listenNestedElement = (element: Element) => {
-        this.listener.listen(element.events, 'changePosition', this.recomputeSelfBounds);
-        this.listener.listen(element.events, 'changeSize', this.recomputeSelfBounds);
-    }
-
     private recomputeSelfBounds = () => {
         if (this.isApplyingParentMove) { return; }
 
-        const {element} = this.context.ontodiaElementContext;
+        const {element} = this.context.ontodiaElement;
         const {x: offsetX, y: offsetY, width: paperWidth, height: paperHeight} = this.getContentFittingBox();
 
         if (this.isNestedElementMoving) {
@@ -173,52 +161,6 @@ export class EmbeddedLayer extends React.Component<{}, State> {
         this.setState({offsetX, offsetY, paperWidth, paperHeight}, () => element.redraw());
     }
 
-    private forceLayout = () => {
-        const {view} = this.context.ontodiaElementContext;
-        const elements = this.getNestedElements();
-
-        const nodeById: { [id: string]: LayoutNode } = {};
-        const nodes: LayoutNode[] = [];
-        for (const element of elements) {
-            const size = element.size;
-            const position = element.position;
-
-            const node: LayoutNode = {
-                id: element.id,
-                x: position.x,
-                y: position.y,
-                width: size.width,
-                height: size.height,
-            };
-            nodeById[element.id] = node;
-            nodes.push(node);
-        }
-
-        const links: LayoutLink[] = [];
-        for (const link of view.model.links) {
-            if (!view.model.isSourceAndTargetVisible(link)) { continue; }
-
-            const source = view.model.sourceOf(link);
-            const target = view.model.targetOf(link);
-
-            const sourceNode = nodeById[source.id];
-            const targetNode = nodeById[target.id];
-
-            if (!sourceNode || !targetNode) { continue; }
-
-            links.push({source: sourceNode, target: targetNode});
-        }
-
-        forceLayout({nodes, links, preferredLinkLength: 200});
-        padded(nodes, {x: 10, y: 10}, () => removeOverlaps(nodes));
-        translateToPositiveQuadrant({nodes, padding: {x: 0, y: 0}});
-
-        for (const node of nodes) {
-            const element = view.model.getElement(node.id);
-            element.setPosition({x: node.x, y: node.y});
-        }
-    }
-
     private onPaperPointerDown = (e: React.MouseEvent<HTMLElement>, cell: Cell | undefined) => {
         if (e.button !== 0 /* left mouse button */) {
             return;
@@ -230,19 +172,10 @@ export class EmbeddedLayer extends React.Component<{}, State> {
         }
     }
 
-    private getParentElement = (layer: HTMLElement): HTMLElement => {
-        const parent = layer.parentElement;
-
-        if (parent.hasAttribute('data-element-id')) {
-            return parent;
-        }
-
-        return this.getParentElement(parent);
-    }
-
-    private calculateOffset = (layer: HTMLElement): { left: number; top: number; } => {
-        const {scale} = this.context.ontodiaElementContext;
-        const parent = this.getParentElement(layer);
+    private calculateOffset(layer: HTMLElement): { left: number; top: number } {
+        const {paperArea} = this.context.ontodiaPaperArea;
+        const scale = paperArea.getScale();
+        const parent = findParentElement(layer);
         const {left, top} = layer.getBoundingClientRect();
         const {left: parentLeft, top: parentTop} = parent.getBoundingClientRect();
 
@@ -259,28 +192,39 @@ export class EmbeddedLayer extends React.Component<{}, State> {
     }
 
     render() {
-        const {view, element, scale} = this.context.ontodiaElementContext;
+        const {element} = this.context.ontodiaElement;
+        const {view} = this.context.ontodiaPaperArea;
         const {paperWidth, paperHeight, offsetX, offsetY} = this.state;
 
-        const style = {
-            position: 'absolute', left: -offsetX, top: -offsetY,
+        const paperTransform: PaperTransform = {
+            width: paperWidth,
+            height: paperHeight,
+            originX: -offsetX,
+            originY: -offsetY,
+            scale: 1,
+            paddingX: 0,
+            paddingY: 0,
         };
 
         return (
             <div className='ontodia-embedded-layer' ref={this.onLayerInit}>
                 <Paper view={view}
-                    width={paperWidth}
-                    height={paperHeight}
-                    originX={-offsetX}
-                    originY={-offsetY}
-                    scale={1}
-                    paddingX={0}
-                    paddingY={0}
+                    paperTransform={paperTransform}
                     onPointerDown={this.onPaperPointerDown}
                     group={element.id}>
-                    <ElementLayer view={view} scale={scale} group={element.id} style={style} />
                 </Paper>
             </div>
         );
+    }
+}
+
+function findParentElement(layer: HTMLElement): HTMLElement {
+    const parent = layer.parentElement;
+    if (!parent) {
+        throw new Error('Cannot find parent diagram element for EmbeddedLayer');
+    } else if (parent.hasAttribute('data-element-id')) {
+        return parent;
+    } else {
+        return findParentElement(parent);
     }
 }
