@@ -4,6 +4,25 @@ export interface Vector {
     readonly x: number;
     readonly y: number;
 }
+export namespace Vector {
+    export function equals(a: Vector, b: Vector) {
+        return a.x === b.x && a.y === b.y;
+    }
+    export function length({x, y}: Vector) {
+        return Math.sqrt(x * x + y * y);
+    }
+    export function normalize({x, y}: Vector) {
+        if (x === 0 && y === 0) { return {x, y}; }
+        const inverseLength = 1 / Math.sqrt(x * x + y * y);
+        return {x: x * inverseLength, y: y * inverseLength};
+    }
+    export function dot({x: x1, y: y1}: Vector, {x: x2, y: y2}: Vector): number {
+        return x1 * x2 + y1 * y2;
+    }
+    export function cross2D({x: x1, y: y1}: Vector, {x: x2, y: y2}: Vector) {
+        return x1 * y2 - y1 * x2;
+    }
+}
 
 export interface Size {
     readonly width: number;
@@ -16,33 +35,16 @@ export interface Rect {
     readonly width: number;
     readonly height: number;
 }
+export namespace Rect {
+    export function center({x, y, width, height}: Rect) {
+        return {x: x + width / 2, y: y + height / 2};
+    }
+}
 
 export function boundsOf(element: DiagramElement): Rect {
     const {x, y} = element.position;
     const {width, height} = element.size;
     return {x, y, width, height};
-}
-
-function centerOfRectangle({x, y, width, height}: Rect): Vector {
-    return {x: x + width / 2, y: y + height / 2};
-}
-
-function length({x, y}: Vector) {
-    return Math.sqrt(x * x + y * y);
-}
-
-export function normalize({x, y}: Vector) {
-    if (x === 0 && y === 0) { return {x, y}; }
-    const inverseLength = 1 / Math.sqrt(x * x + y * y);
-    return {x: x * inverseLength, y: y * inverseLength};
-}
-
-function dot({x: x1, y: y1}: Vector, {x: x2, y: y2}: Vector): number {
-    return x1 * x2 + y1 * y2;
-}
-
-function cross2D({x: x1, y: y1}: Vector, {x: x2, y: y2}: Vector) {
-    return x1 * y2 - y1 * x2;
 }
 
 function intersectRayFromRectangleCenter(sourceRect: Rect, rayTarget: Vector) {
@@ -61,15 +63,15 @@ function intersectRayFromRectangleCenter(sourceRect: Rect, rayTarget: Vector) {
         return center;
     }
 
-    const direction = normalize({
+    const direction = Vector.normalize({
         x: rayTarget.x - center.x,
         y: rayTarget.y - center.y,
     });
 
     const rightDirection = {x: Math.abs(direction.x), y: direction.y};
     const isHorizontal =
-        cross2D({x: halfWidth, y: -halfHeight}, rightDirection) > 0 &&
-        cross2D({x: halfWidth, y: halfHeight}, rightDirection) < 0;
+        Vector.cross2D({x: halfWidth, y: -halfHeight}, rightDirection) > 0 &&
+        Vector.cross2D({x: halfWidth, y: halfHeight}, rightDirection) < 0;
 
     if (isHorizontal) {
         return {
@@ -84,6 +86,19 @@ function intersectRayFromRectangleCenter(sourceRect: Rect, rayTarget: Vector) {
     }
 }
 
+export function isPolylineEqual(left: ReadonlyArray<Vector>, right: ReadonlyArray<Vector>) {
+    if (left === right) { return true; }
+    if (left.length !== right.length) { return false; }
+    for (let i = 0; i < left.length; i++) {
+        const a = left[i];
+        const b = right[i];
+        if (!(a.x === b.x && a.y === b.y)) {
+            return false;
+        }
+    }
+    return true;
+}
+
 export function computePolyline(
     source: DiagramElement,
     target: DiagramElement,
@@ -92,16 +107,16 @@ export function computePolyline(
     const sourceRect = boundsOf(source);
     const targetRect = boundsOf(target);
     const startPoint = intersectRayFromRectangleCenter(
-        sourceRect, vertices.length > 0 ? vertices[0] : centerOfRectangle(targetRect));
+        sourceRect, vertices.length > 0 ? vertices[0] : Rect.center(targetRect));
     const endPoint = intersectRayFromRectangleCenter(
-        targetRect, vertices.length > 0 ? vertices[vertices.length - 1] : centerOfRectangle(sourceRect));
+        targetRect, vertices.length > 0 ? vertices[vertices.length - 1] : Rect.center(sourceRect));
     return [startPoint, ...vertices, endPoint];
 }
 
 export function computePolylineLength(polyline: ReadonlyArray<Vector>): number {
     let previous: Vector;
     return polyline.reduce((acc, point) => {
-        const segmentLength = previous ? length({x: point.x - previous.x, y: point.y - previous.y}) : 0;
+        const segmentLength = previous ? Vector.length({x: point.x - previous.x, y: point.y - previous.y}) : 0;
         previous = point;
         return acc + segmentLength;
     }, 0);
@@ -119,7 +134,7 @@ export function getPointAlongPolyline(polyline: ReadonlyArray<Vector>, offset: n
         const previous = polyline[i - 1];
         const point = polyline[i];
         const segment = {x: point.x - previous.x, y: point.y - previous.y};
-        const segmentLength = length(segment);
+        const segmentLength = Vector.length(segment);
         const newOffset = currentOffset + segmentLength;
         if (offset < newOffset) {
             const leftover = (offset - currentOffset) / segmentLength;
@@ -144,18 +159,34 @@ export function findNearestSegmentIndex(polyline: ReadonlyArray<Vector>, locatio
 
         const target = {x: location.x - pivot.x, y: location.y - pivot.y};
         const segment = {x: next.x - pivot.x, y: next.y - pivot.y};
-        const segmentLength = length(segment);
+        const segmentLength = Vector.length(segment);
 
-        const projectionToSegment = dot(target, segment) / segmentLength;
+        const projectionToSegment = Vector.dot(target, segment) / segmentLength;
         if (projectionToSegment < 0 || projectionToSegment > segmentLength) {
             continue;
         }
 
-        const distanceToSegment = Math.abs(cross2D(target, segment)) / segmentLength;
+        const distanceToSegment = Math.abs(Vector.cross2D(target, segment)) / segmentLength;
         if (distanceToSegment < minDistance) {
             minDistance = distanceToSegment;
             foundIndex = i;
         }
     }
     return foundIndex;
+}
+
+export function computeGrouping(elements: ReadonlyArray<DiagramElement>): Map<string, DiagramElement[]> {
+    const grouping = new Map<string, DiagramElement[]>();
+    for (const element of elements) {
+        const group = element.group;
+        if (typeof group === 'string') {
+            let children = grouping.get(group);
+            if (!children) {
+                children = [];
+                grouping.set(group, children);
+            }
+            children.push(element);
+        }
+    }
+    return grouping;
 }
