@@ -1,31 +1,27 @@
 import * as React from 'react';
 import { Component, SVGAttributes, CSSProperties } from 'react';
 
-import { Element as DiagramElement, Link as DiagramLink } from './elements';
+import { Cell, Element as DiagramElement, Link as DiagramLink, LinkVertex } from './elements';
+import { ElementLayer } from './elementLayer';
+import { Vector } from './geometry';
 import { LinkLayer, LinkMarkers } from './linkLayer';
 import { DiagramModel } from './model';
 import { DiagramView } from './view';
 
 export interface PaperProps {
     view: DiagramView;
-    width: number;
-    height: number;
-    originX: number;
-    originY: number;
-    scale: number;
-    paddingX: number;
-    paddingY: number;
+    paperTransform: PaperTransform;
     onPointerDown?: (e: React.MouseEvent<HTMLElement>, cell: Cell | undefined) => void;
     group?: string;
 }
 
-export type Cell = DiagramElement | DiagramLink | LinkVertex;
-
 const CLASS_NAME = 'ontodia-paper';
 
-export class Paper extends Component<PaperProps, void> {
+export class Paper extends Component<PaperProps, {}> {
     render() {
-        const {width, height, originX, originY, scale, paddingX, paddingY} = this.props;
+        const {view, group, paperTransform} = this.props;
+        const {width, height, originX, originY, scale, paddingX, paddingY} = paperTransform;
+
         const scaledWidth = width * scale;
         const scaledHeight = height * scale;
         // using padding instead of margin in combination with setting width and height
@@ -39,16 +35,20 @@ export class Paper extends Component<PaperProps, void> {
             paddingRight: paddingX,
             paddingBottom: paddingY,
         };
+        const htmlTransformStyle: React.CSSProperties = {
+            position: 'absolute', left: 0, top: 0,
+            transform: `scale(${scale},${scale})translate(${originX}px,${originY}px)`,
+        };
+
         return (
             <div className={CLASS_NAME} style={style} onMouseDown={this.onMouseDown}>
-                <svg className={`${CLASS_NAME}__canvas`}
-                    width={scaledWidth} height={scaledHeight}
-                    style={{overflow: 'visible'}}>
-                    <LinkMarkers view={this.props.view} />
-                    <g transform={`scale(${scale},${scale})translate(${originX},${originY})`}>
-                        <LinkLayer view={this.props.view} group={this.props.group} />
-                    </g>
-                </svg>
+                <TransformedSvgCanvas className={`${CLASS_NAME}__canvas`}
+                    style={{overflow: 'visible'}}
+                    paperTransform={paperTransform}>
+                    <LinkMarkers view={view} />
+                    <LinkLayer view={view} links={view.model.links} group={group} />
+                </TransformedSvgCanvas>
+                <ElementLayer view={view} group={group} style={htmlTransformStyle} />
                 {this.props.children}
             </div>
         );
@@ -64,27 +64,16 @@ export class Paper extends Component<PaperProps, void> {
     }
 }
 
-export interface LinkVertex {
-    link: DiagramLink;
-    vertexIndex: number;
-}
-
-export function isLinkVertex(cell: Cell | undefined): cell is LinkVertex {
-    return cell && typeof cell === 'object'
-        && 'link' in cell
-        && 'vertexIndex' in cell;
-}
-
 function findCell(bottom: Element, top: Element, model: DiagramModel): Cell | undefined {
     let target: Node = bottom;
-    let vertexIndex: number | undefined = undefined;
+    let vertexIndex: number | undefined;
     while (true) {
         if (target instanceof Element) {
             if (target.hasAttribute('data-element-id')) {
                 return model.getElement(target.getAttribute('data-element-id'));
             } else if (target.hasAttribute('data-link-id')) {
                 const link = model.getLinkById(target.getAttribute('data-link-id'));
-                return typeof vertexIndex === 'number' ? {link, vertexIndex} : link;
+                return typeof vertexIndex === 'number' ? new LinkVertex(link, vertexIndex) : link;
             } else if (target.hasAttribute('data-vertex')) {
                 vertexIndex = Number(target.getAttribute('data-vertex'));
             }
@@ -93,4 +82,43 @@ function findCell(bottom: Element, top: Element, model: DiagramModel): Cell | un
         target = target.parentNode;
     }
     return undefined;
+}
+
+export interface PaperTransform {
+    width: number;
+    height: number;
+    originX: number;
+    originY: number;
+    scale: number;
+    paddingX: number;
+    paddingY: number;
+}
+
+export interface TransformedSvgCanvasProps extends React.HTMLProps<SVGSVGElement> {
+    paperTransform: PaperTransform;
+}
+
+export class TransformedSvgCanvas extends Component<TransformedSvgCanvasProps, {}> {
+    private static readonly SVG_STYLE: CSSProperties = {
+        position: 'absolute',
+        top: 0,
+        left: 0,
+    };
+    render() {
+        const {paperTransform, style, children, ...otherProps} = this.props;
+        const {width, height, originX, originY, scale, paddingX, paddingY} = paperTransform;
+        const scaledWidth = width * scale;
+        const scaledHeight = height * scale;
+        let svgStyle = TransformedSvgCanvas.SVG_STYLE;
+        if (style) {
+            svgStyle = {...svgStyle, ...style};
+        }
+        return (
+            <svg width={scaledWidth} height={scaledHeight} style={svgStyle} {...otherProps}>
+                <g transform={`scale(${scale},${scale})translate(${originX},${originY})`}>
+                    {children}
+                </g>
+            </svg>
+        );
+    }
 }
