@@ -1,52 +1,38 @@
 import { hcl } from 'd3-color';
 import { defaultsDeep, cloneDeep } from 'lodash';
-import { ReactElement, createElement } from 'react';
+import { ReactElement } from 'react';
 
 import {
-    TypeStyleResolver,
-    LinkTemplateResolver,
-    TemplateResolver,
-    CustomTypeStyle,
-    ElementTemplate,
-    LinkTemplate,
-    LinkMarkerStyle,
-    LinkRouter,
+    LinkRouter, TypeStyleResolver, LinkTemplateResolver, TemplateResolver,
+    CustomTypeStyle, ElementTemplate, LinkTemplate, RoutedLink, RoutedLinks,
 } from '../customization/props';
 import { DefaultTypeStyleBundle } from '../customization/defaultTypeStyles';
 import { DefaultLinkTemplateBundle } from '../customization/defaultLinkStyles';
 import { StandardTemplate, DefaultTemplateBundle } from '../customization/templates';
 
+import { ElementModel, LocalizedString, ElementTypeIri, LinkTypeIri } from '../data/model';
+import { isEncodedBlank } from '../data/sparql/blankNodes';
+import { hashFnv32a, uri2name } from '../data/utils';
+
 import { Events, EventSource, EventObserver, PropertyChange } from '../viewUtils/events';
-import { ConnectionsMenu, PropertySuggestionHandler } from '../widgets/connectionsMenu';
-import { Halo } from '../widgets/halo';
 
-import { Dictionary, ElementModel, LocalizedString } from '../data/model';
-import { hashFnv32a } from '../data/utils';
+import { Element, FatLinkType, FatClassModel } from './elements';
+import { DiagramModel, chooseLocalizedText } from './model';
+import { AsyncModel } from '../editor/asyncModel';
 
-import { Element, Link, FatLinkType, FatClassModel, linkMarkerKey } from './elements';
-import { Size, boundsOf } from './geometry';
-import { DiagramModel, chooseLocalizedText, uri2name } from './model';
-import { isLinkVertex } from './paper';
-import { PaperArea, PointerUpEvent } from './paperArea';
 
-export interface DiagramViewOptions {
+import { DefaultLinkRouter } from './linkRouter';
+
+export interface ViewOptions {
     typeStyleResolvers?: TypeStyleResolver[];
     linkTemplateResolvers?: LinkTemplateResolver[];
     templatesResolvers?: TemplateResolver[];
-    disableDefaultHalo?: boolean;
     linkRouter?: LinkRouter;
-    suggestProperties?: PropertySuggestionHandler;
     onIriClick?: (iri: string, element: Element, event: React.MouseEvent<any>) => void;
-    groupBy?: GroupBy[];
-}
-
-export interface GroupBy {
-    linkType: string;
-    linkDirection: 'in' | 'out';
 }
 
 export interface TypeStyle {
-    color: { h: number; c: number; l: number; };
+    color: { h: number; c: number; l: number };
     icon?: string;
 }
 
@@ -55,23 +41,27 @@ export enum RenderingLayer {
     ElementSize,
     PaperArea,
     Link,
+    Editor,
 
     FirstToUpdate = Element,
-    LastToUpdate = Link,
+    LastToUpdate = Editor,
 }
 
 export interface DiagramViewEvents {
     changeLanguage: PropertyChange<DiagramView, string>;
-    changeSelection: PropertyChange<DiagramView, ReadonlyArray<Element>>;
-    changeLinkTemplates: { source: DiagramView };
-    toggleNavigationMenu: { isOpened: boolean };
+    changeLinkTemplates: {};
     syncUpdate: { layer: RenderingLayer };
     updateWidgets: UpdateWidgetsEvent;
-    dispose: { source: DiagramView };
+    dispose: {};
 }
 
 export interface UpdateWidgetsEvent {
-    widgets: { [key: string]: ReactElement<any> };
+    widgets: { [key: string]: WidgetDescription };
+}
+
+export interface WidgetDescription {
+    element: ReactElement<any>;
+    pinnedToScreen: boolean;
 }
 
 export class DiagramView {
@@ -87,15 +77,16 @@ export class DiagramView {
     private linkTemplateResolvers: LinkTemplateResolver[];
     private templatesResolvers: TemplateResolver[];
 
-    private connectionsMenuTarget: Element | undefined;
-
     private _language = 'en';
-    private _selection: ReadonlyArray<Element> = [];
-    private linkTemplates: { [linkTypeId: string]: LinkTemplate } = {};
+
+    private linkTemplates = new Map<LinkTypeIri, LinkTemplate>();
+
+    private router: LinkRouter;
+    private routings: RoutedLinks;
 
     constructor(
-        public readonly model: DiagramModel,
-        public readonly options: DiagramViewOptions = {},
+        public readonly model: AsyncModel,
+        public readonly options: ViewOptions = {},
     ) {
         this.typeStyleResolvers = options.typeStyleResolvers
             ? options.typeStyleResolvers : DefaultTypeStyleBundle;
@@ -105,14 +96,33 @@ export class DiagramView {
 
         this.templatesResolvers = options.templatesResolvers
             ? options.templatesResolvers : DefaultTemplateBundle;
+
+        this.initRouting();
     }
 
-    get selection() { return this._selection; }
-    setSelection(value: ReadonlyArray<Element>) {
-        const previous = this._selection;
-        if (previous === value) { return; }
-        this._selection = value;
-        this.source.trigger('changeSelection', {source: this, previous});
+    private initRouting() {
+        this.router = this.options.linkRouter || new DefaultLinkRouter();
+        this.updateRoutings();
+
+        this.listener.listen(this.model.events, 'changeCells', () =>  this.updateRoutings());
+        this.listener.listen(this.model.events, 'linkEvent', ({key, data}) => {
+            if (data.changeVertices) {
+                this.updateRoutings();
+            }
+        });
+        this.listener.listen(this.model.events, 'elementEvent', ({key, data}) => {
+            if (data.changePosition || data.changeSize) {
+                this.updateRoutings();
+            }
+        });
+    }
+
+    private updateRoutings() {
+        this.routings = this.router.route(this.model);
+    }
+
+    getRouting(linkId: string): RoutedLink {
+        return this.routings[linkId];
     }
 
     getLanguage(): string { return this._language; }
@@ -126,69 +136,13 @@ export class DiagramView {
         this.source.trigger('changeLanguage', {source: this, previous});
     }
 
-    getLinkTemplates(): { readonly [linkTypeId: string]: LinkTemplate } {
+    getLinkTemplates(): ReadonlyMap<LinkTypeIri, LinkTemplate> {
         return this.linkTemplates;
-    }
-
-    cancelSelection() {
-        this.setSelection([]);
     }
 
     performSyncUpdate() {
         for (let layer = RenderingLayer.FirstToUpdate; layer <= RenderingLayer.LastToUpdate; layer++) {
             this.source.trigger('syncUpdate', {layer});
-        }
-    }
-
-    _initializePaperComponents(paperArea: PaperArea) {
-        this.listener.listen(paperArea.events, 'pointerUp', e => this.onPaperPointerUp(e));
-
-        if (!this.options.disableDefaultHalo) {
-            this.configureHalo();
-            document.addEventListener('keyup', this.onKeyUp);
-            this.listener.listen(this.events, 'dispose', () => {
-                document.removeEventListener('keyup', this.onKeyUp);
-            });
-        }
-    }
-
-    private onKeyUp = (e: KeyboardEvent) => {
-        const DELETE_KEY_CODE = 46;
-        if (e.keyCode === DELETE_KEY_CODE &&
-            document.activeElement.localName !== 'input'
-        ) {
-            this.removeSelectedElements();
-        }
-    }
-
-    removeSelectedElements() {
-        const elementsToRemove = this.selection;
-        if (elementsToRemove.length === 0) { return; }
-
-        this.cancelSelection();
-        this.model.initBatchCommand();
-        for (const element of elementsToRemove) {
-            this.model.removeElement(element.id);
-        }
-        this.setSelection([]);
-        this.model.storeBatchCommand();
-    }
-
-    private onPaperPointerUp(event: PointerUpEvent) {
-        if (this.options.disableDefaultHalo) { return; }
-        const {sourceEvent, target, triggerAsClick} = event;
-        
-        if (sourceEvent.ctrlKey || sourceEvent.shiftKey || sourceEvent.metaKey) { return; }
-        
-        if (target instanceof Element) {
-            this.setSelection([target]);
-            target.focus();
-        } else if (!target && triggerAsClick) {
-            this.setSelection([]);
-            this.hideNavigationMenu();
-            if (document.activeElement) {
-                (document.activeElement as HTMLElement).blur();
-            }
         }
     }
 
@@ -201,132 +155,14 @@ export class DiagramView {
         }
     }
 
-    private configureHalo() {
-        if (this.options.disableDefaultHalo) { return; }
-
-        const renderDefaultHalo = () => {
-            const selectedElement = this.selection.length === 1 ? this.selection[0] : undefined;
-            const halo = createElement(Halo, {
-                diagramView: this,
-                target: selectedElement,
-                onDelete: () => this.removeSelectedElements(),
-                onExpand: () => {
-                    selectedElement.setExpanded(!selectedElement.isExpanded);
-                },
-                navigationMenuOpened: Boolean(this.connectionsMenuTarget),
-                onToggleNavigationMenu: () => {
-                    if (this.connectionsMenuTarget) {
-                        this.hideNavigationMenu();
-                    } else {
-                        this.showNavigationMenu(selectedElement);
-                    }
-                    renderDefaultHalo();
-                },
-                onAddToFilter: () => selectedElement.addToFilter(),
-            });
-            this.source.trigger('updateWidgets', {widgets: {halo}});
-        };
-
-        this.listener.listen(this.events, 'changeSelection', () => {
-            const selected = this.selection.length === 1 ? this.selection[0] : undefined;
-            if (this.connectionsMenuTarget && selected !== this.connectionsMenuTarget) {
-                this.hideNavigationMenu();
-            }
-            renderDefaultHalo();
-        });
-
-        this.listener.listen(this.events, 'toggleNavigationMenu', ({isOpened}) => {
-            renderDefaultHalo();
-        });
-
-        renderDefaultHalo();
-    }
-
-    setPaperWidget(widget: { key: string; widget: ReactElement<any>; }) {
-        const widgets = {[widget.key]: widget.widget};
+    setPaperWidget(widget: {
+        key: string;
+        widget: ReactElement<any> | undefined;
+        pinnedToScreen?: boolean;
+    }) {
+        const {key, widget: element, pinnedToScreen} = widget;
+        const widgets = {[widget.key]: element ? {element, pinnedToScreen} : undefined};
         this.source.trigger('updateWidgets', {widgets});
-    }
-
-    showNavigationMenu(target: Element) {
-        const connectionsMenu = createElement(ConnectionsMenu, {
-            view: this,
-            target,
-            onClose: () => this.hideNavigationMenu(),
-            suggestProperties: this.options.suggestProperties,
-        });
-        this.connectionsMenuTarget = target;
-        this.source.trigger('updateWidgets', {widgets: {connectionsMenu}});
-        this.source.trigger('toggleNavigationMenu', {isOpened: false});
-    }
-
-    hideNavigationMenu() {
-        if (this.connectionsMenuTarget) {
-            this.connectionsMenuTarget = undefined;
-            this.source.trigger('updateWidgets', {widgets: {connectionsMenu: undefined}});
-            this.source.trigger('toggleNavigationMenu', {isOpened: false});
-        }
-    }
-
-    onDragDrop(e: DragEvent, paperPosition: { x: number; y: number; }) {
-        e.preventDefault();
-        let elementIds: string[];
-        try {
-            elementIds = JSON.parse(e.dataTransfer.getData('application/x-ontodia-elements'));
-        } catch (ex) {
-            try {
-                elementIds = JSON.parse(e.dataTransfer.getData('text')); // IE fix
-            } catch (ex) {
-                const draggedUri = e.dataTransfer.getData('text/uri-list');
-                // element dragged from the class tree has URI of the form:
-                // <window.location without hash>#<class URI>
-                const uriFromTreePrefix = window.location.href.split('#')[0] + '#';
-                const uri = draggedUri.indexOf(uriFromTreePrefix) === 0
-                    ? draggedUri.substring(uriFromTreePrefix.length) : draggedUri;
-                elementIds = [uri];
-            }
-        }
-        if (!elementIds || elementIds.length === 0) { return; }
-
-        this.model.initBatchCommand();
-
-        let elementsToSelect: Element[] = [];
-
-        let totalXOffset = 0;
-        let {x, y} = paperPosition;
-        for (const elementId of elementIds) {
-            const center = elementIds.length === 1;
-            const {element, size} = this.createElementAt(elementId, {x: x + totalXOffset, y, center});
-            totalXOffset += size.width + 20;
-
-            elementsToSelect.push(element);
-            element.focus();
-        }
-
-        this.model.requestElementData(elementsToSelect);
-        this.model.requestLinksOfType();
-        this.setSelection(elementsToSelect);
-
-        this.model.storeBatchCommand();
-    }
-
-    private createElementAt(
-        elementId: string,
-        position: { x: number; y: number; center?: boolean; }
-    ): { element: Element, size: Size } {
-        const element = this.model.createElement(elementId);
-
-        let {x, y} = position;
-        let {width, height} = boundsOf(element);
-        if (width === 0) { width = 100; }
-        if (height === 0) { height = 50; }
-
-        if (position.center) {
-            x -= width / 2;
-            y -= height / 2;
-        }
-        element.setPosition({x, y});
-
-        return {element, size: {width, height}};
     }
 
     /**
@@ -338,8 +174,8 @@ export class DiagramView {
     }
 
     public getElementTypeString(elementModel: ElementModel): string {
-        return elementModel.types.map((typeId: string) => {
-            const type = this.model.getClassesById(typeId);
+        return elementModel.types.map(typeId => {
+            const type = this.model.createClass(typeId);
             return this.getElementTypeLabel(type).text;
         }).sort().join(', ');
     }
@@ -349,13 +185,13 @@ export class DiagramView {
         return label ? label : { text: uri2name(type.id), lang: '' };
     }
 
-    public getLinkLabel(linkTypeId: string): LocalizedString {
-        const type = this.model.getLinkType(linkTypeId);
+    public getLinkLabel(linkTypeId: LinkTypeIri): LocalizedString {
+        const type = this.model.createLinkType(linkTypeId);
         const label = type ? this.getLocalizedText(type.label) : null;
         return label ? label : { text: uri2name(linkTypeId), lang: '' };
     }
 
-    public getTypeStyle(types: string[]): TypeStyle {
+    public getTypeStyle(types: ElementTypeIri[]): TypeStyle {
         types.sort();
 
         let customStyle: CustomTypeStyle;
@@ -368,7 +204,7 @@ export class DiagramView {
         }
 
         const icon = customStyle ? customStyle.icon : undefined;
-        let color: { h: number; c: number; l: number; };
+        let color: { h: number; c: number; l: number };
         if (customStyle && customStyle.color) {
             color = hcl(customStyle.color);
         } else {
@@ -376,6 +212,13 @@ export class DiagramView {
             color = {h: hue, c: 40, l: 75};
         }
         return {icon, color};
+    }
+
+    formatIri(iri: string): string {
+        if (isEncodedBlank(iri)) {
+            return '(blank node)';
+        }
+        return `<${iri}>`;
     }
 
     public registerElementStyleResolver(resolver: TypeStyleResolver): TypeStyleResolver {
@@ -392,7 +235,7 @@ export class DiagramView {
         }
     }
 
-    public getElementTemplate(types: string[]): ElementTemplate {
+    public getElementTemplate(types: ElementTypeIri[]): ElementTemplate {
         for (const resolver of this.templatesResolvers) {
             const result = resolver(types);
             if (result) {
@@ -417,7 +260,7 @@ export class DiagramView {
     }
 
     createLinkTemplate(linkType: FatLinkType): LinkTemplate {
-        const existingTemplate = this.linkTemplates[linkType.id];
+        const existingTemplate = this.linkTemplates.get(linkType.id);
         if (existingTemplate) {
             return existingTemplate;
         }
@@ -431,9 +274,9 @@ export class DiagramView {
             }
         }
 
-        fillLinkTemplateDefaults(template, this.model);
-        this.linkTemplates[linkType.id] = template;
-        this.source.trigger('changeLinkTemplates', {source: this});
+        fillLinkTemplateDefaults(template);
+        this.linkTemplates.set(linkType.id, template);
+        this.source.trigger('changeLinkTemplates', {});
         return template;
     }
 
@@ -449,27 +292,13 @@ export class DiagramView {
 
     dispose() {
         if (this.disposed) { return; }
-        this.source.trigger('dispose', {source: this});
+        this.source.trigger('dispose', {});
         this.listener.stopListening();
         this.disposed = true;
     }
-
-    loadEmbeddedElements = (elementId: string): Promise<Dictionary<ElementModel>> => {
-        const elements = this.options.groupBy.map(groupBy =>
-            this.model.dataProvider.linkElements({
-                elementId,
-                linkId: groupBy.linkType,
-                offset: 0,
-                direction: groupBy.linkDirection,
-            })
-        );
-        return Promise.all(elements).then(res =>
-            res.reduce((memo, current) => Object.assign(memo, current), {})
-        );
-    }
 }
 
-function getHueFromClasses(classes: string[], seed?: number): number {
+function getHueFromClasses(classes: ReadonlyArray<ElementTypeIri>, seed?: number): number {
     let hash = seed;
     for (const name of classes) {
         hash = hashFnv32a(name, hash);
@@ -478,7 +307,7 @@ function getHueFromClasses(classes: string[], seed?: number): number {
     return 360 * ((hash === undefined ? 0 : hash) / MAX_INT32);
 }
 
-function fillLinkTemplateDefaults(template: LinkTemplate, model: DiagramModel) {
+function fillLinkTemplateDefaults(template: LinkTemplate) {
     const defaults: Partial<LinkTemplate> = {
         markerTarget: {d: 'M0,0 L0,8 L9,4 z', width: 9, height: 8, fill: 'black'},
     };
@@ -487,5 +316,3 @@ function fillLinkTemplateDefaults(template: LinkTemplate, model: DiagramModel) {
         template.renderLink = () => ({});
     }
 }
-
-export default DiagramView;
