@@ -2,7 +2,7 @@ import * as React from 'react';
 
 export interface Props {
     size?: number;
-    position?: { x: number; y: number; };
+    position?: { x: number; y: number };
     maxWidth?: number;
     statusText?: string;
     errorOccured?: boolean;
@@ -10,9 +10,9 @@ export interface Props {
 
 const CLASS_NAME = 'ontodia-spinner';
 
-export class Spinner extends React.Component<Props, void> {
+export class Spinner extends React.Component<Props, {}> {
     render() {
-        let {position = {x: 0, y: 0}, size = 50, statusText, errorOccured} = this.props;
+        const {position = {x: 0, y: 0}, size = 50, statusText, errorOccured} = this.props;
 
         const textLeftMargin = 5;
         const pathGeometry = 'm3.47,-19.7 a20,20 0 1,1 -6.95,0 m0,0 l-6,5 m6,-5 l-8,-0' +
@@ -27,5 +27,17 @@ export class Spinner extends React.Component<Props, void> {
             </g>
             <text style={{dominantBaseline: 'middle'}} x={size / 2 + textLeftMargin}>{statusText}</text>
         </g>;
+    }
+}
+
+export class HtmlSpinner extends React.Component<{ width: number; height: number }, {}> {
+    render() {
+        const {width, height} = this.props;
+        const size = Math.min(width, height);
+        return (
+            <svg width={width} height={height}>
+                <Spinner size={size} position={{x: width / 2, y: height / 2}} />
+            </svg>
+        );
     }
 }
