@@ -1,12 +1,11 @@
-var webpack = require('webpack');
 var path = require('path');
-
-var CommonsChunkPlugin = require("webpack/lib/optimize/CommonsChunkPlugin");
 var HtmlWebpackPlugin = require('html-webpack-plugin');
 
-var SUPPORT_IE = Boolean(process.env.SUPPORT_IE);
-
-var npmDir = path.join(__dirname, 'node_modules');
+var SUPPORT_IE = process.env.SUPPORT_IE;
+var SPARQL_ENDPOINT = process.env.SPARQL_ENDPOINT;
+var WIKIDATA_ENDPOINT = process.env.WIKIDATA_ENDPOINT;
+var LOD_PROXY = process.env.LOD_PROXY;
+var PROP_SUGGEST = process.env.PROP_SUGGEST;
 
 var aliases = {};
 if (!SUPPORT_IE) {
@@ -15,143 +14,165 @@ if (!SUPPORT_IE) {
     aliases['es6-promise/auto'] = emptyModule;
 }
 
+var examplesDir = path.join(__dirname, 'src', 'examples');
+var htmlTemplatePath = path.join(__dirname, 'src', 'examples', 'template.ejs');
+
 module.exports = {
+    mode: 'development',
     entry: {
-        rdf: path.join(__dirname, 'src', 'examples', 'rdf.ts'),
-        demo: path.join(__dirname, 'src', 'examples', 'demo.ts'),
-        sparql: path.join(__dirname, 'src', 'examples', 'sparql.ts'),
-        dbpedia: path.join(__dirname, 'src', 'examples', 'dbpedia.ts'),
-        sparqlNoStats: path.join(__dirname, 'src', 'examples', 'sparqlNoStats.ts'),
-        sparqlConstruct: path.join(__dirname, 'src', 'examples', 'sparqlConstruct.ts'),
-        sparqlRDFGraph: path.join(__dirname, 'src', 'examples', 'sparqlRDFGraph.ts'),
-        sparqlTurtleGraph: path.join(__dirname, 'src', 'examples', 'sparqlTurtleGraph.ts'),
-        styleCustomization: path.join(__dirname, 'src', 'examples', 'styleCustomization.ts'),
-        wikidata: path.join(__dirname, 'src', 'examples', 'wikidata.ts'),
-        composite: path.join(__dirname, 'src', 'examples', 'composite.ts'),
-        wikidataGraph: path.join(__dirname, 'src', 'examples', 'wikidataGraph.ts'),
-        toolbarCustomization: path.join(__dirname, 'src', 'examples', 'toolbarCustomization.tsx'),
+        rdf: path.join(examplesDir, 'rdf.ts'),
+        demo: path.join(examplesDir, 'demo.ts'),
+        sparql: path.join(examplesDir, 'sparql.ts'),
+        dbpedia: path.join(examplesDir, 'dbpedia.ts'),
+        sparqlNoStats: path.join(examplesDir, 'sparqlNoStats.ts'),
+        sparqlConstruct: path.join(examplesDir, 'sparqlConstruct.ts'),
+        sparqlRDFGraph: path.join(examplesDir, 'sparqlRDFGraph.ts'),
+        sparqlTurtleGraph: path.join(examplesDir, 'sparqlTurtleGraph.ts'),
+        styleCustomization: path.join(examplesDir, 'styleCustomization.ts'),
+        wikidata: path.join(examplesDir, 'wikidata.ts'),
+        composite: path.join(examplesDir, 'composite.ts'),
+        wikidataGraph: path.join(examplesDir, 'wikidataGraph.ts'),
+        toolbarCustomization: path.join(examplesDir, 'toolbarCustomization.tsx'),
     },
     resolve: {
         alias: aliases,
-        extensions: ['', '.ts', '.tsx', '.webpack.js', '.web.js', '.js'],
+        extensions: ['.ts', '.tsx', '.js'],
     },
     module: {
-        loaders: [
-            {test: /\.ts$|\.tsx$/, loader: 'ts-loader'},
-            {test: /\.css$/, loader: 'style-loader!css-loader'},
-            {test: /\.scss$/, loader: 'style-loader!css-loader!sass-loader'},
-            {test: /\.jpe?g$/, loader: 'url-loader?mimetype=image/jpeg'},
-            {test: /\.gif$/, loader: 'url-loader?mimetype=image/gif'},
-            {test: /\.png$/, loader: 'url-loader?mimetype=image/png'},
-            {test: /\.svg$/, loader: 'url-loader?mimetype=image/svg+xml'},
-        ],
+        rules: [
+            {test: /\.ts$|\.tsx$/, use: ['ts-loader']},
+            {test: /\.css$/, use: ['style-loader', 'css-loader']},
+            {test: /\.scss$/, use: ['style-loader', 'css-loader', 'sass-loader']},
+            {
+                test: /\.(jpe?g|gif|png|svg)$/,
+                use: [{loader: 'url-loader'}],
+            },
+            {test: /\.ttl$/, use: ['raw-loader']},
+        ]
+    },
+    devtool: 'source-map',
+    optimization: {
+        splitChunks: {
+            cacheGroups: {
+                commons: {
+                    name: 'commons',
+                    chunks: 'initial',
+                    minChunks: 2,
+                }
+            }
+        },
+    },
+    output: {
+        path: path.join(__dirname, 'dist', 'examples'),
+        filename: '[name].bundle.js',
+        chunkFilename: '[id].chunk.js',
+        publicPath: '',
     },
     plugins: [
         new HtmlWebpackPlugin({
             filename: 'rdf.html',
             title: 'Ontodia RDF Demo',
             chunks: ['commons', 'rdf'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             title: 'Ontodia Local Demo',
             chunks: ['commons', 'demo'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'sparql.html',
             title: 'Ontodia SparQL Demo',
             chunks: ['commons', 'sparql'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'dbpedia.html',
             title: 'Ontodia DBPedia SparQL Demo',
             chunks: ['commons', 'dbpedia'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'sparqlNoStats.html',
             title: 'Ontodia SparQL Demo',
             chunks: ['commons', 'sparqlNoStats'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'sparqlConstruct.html',
             title: 'Ontodia SparQL Construct Demo',
             chunks: ['commons', 'sparqlConstruct'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'sparqlRDFGraph.html',
             title: 'Ontodia SparQL RDF Graph Demo',
             chunks: ['commons', 'sparqlRDFGraph'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'sparqlTurtleGraph.html',
             title: 'Ontodia SparQL Turtle Graph Demo',
             chunks: ['commons', 'sparqlTurtleGraph'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'styleCustomization.html',
             title: 'Ontodia Style Customization Demo',
             chunks: ['commons', 'styleCustomization', ],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'wikidata.html',
             title: 'Ontodia Wikidata Demo',
             chunks: ['commons', 'wikidata', ],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'wikidataGraph.html',
             title: 'Ontodia Wikidata with graph Demo',
             chunks: ['commons', 'wikidataGraph', ],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'composite.html',
             title: 'Ontodia composite DP Demo',
             chunks: ['commons', 'composite'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
         new HtmlWebpackPlugin({
             filename: 'toolbarCustomization.html',
             title: 'Ontodia Toolbar Customization Demo',
             chunks: ['commons', 'toolbarCustomization'],
-            template: path.join(__dirname, 'src', 'examples', 'template.ejs'),
+            template: htmlTemplatePath,
         }),
-        new CommonsChunkPlugin('commons', 'commons.chunk.js'),
     ],
-    output: {
-        path: path.join(__dirname, 'dist', 'examples'),
-        filename: '[name].bundle.js',
-        chunkFilename: '[id].chunk.js',
-        publicPath: '/',
-    },
-    devtool: '#source-map',
     devServer: {
+        contentBase: './dist',
         proxy: {
-            '/sparql-endpoint**': {
-                target: process.env.SPARQL_ENDPOINT,
-                pathRewrite: {'/sparql-endpoint' : ''},
+            '/sparql**': {
+                target: SPARQL_ENDPOINT,
+                pathRewrite: {'/sparql' : ''},
+                changeOrigin: true,
+                secure: false,
+            },
+            '/wikidata**': {
+                target: WIKIDATA_ENDPOINT || SPARQL_ENDPOINT,
+                pathRewrite: {'/wikidata' : ''},
                 changeOrigin: true,
                 secure: false,
             },
             '/lod-proxy/**': {
-                target: process.env.LOD_PROXY,
+                target: LOD_PROXY,
                 changeOrigin: true,
                 secure: false,
             },
             '/wikidata-prop-suggest**': {
-                target: 'http://wikidata-prop-suggest.apps.vismart.biz/',
+                target: PROP_SUGGEST,
                 pathRewrite: {'/wikidata-prop-suggest' : ''},
                 changeOrigin: true,
                 secure: false,
             },
         },
-    },
+    }
 };
